Add `AnyCodable` import to swift 5 api template
__Why the change?__

I ran into an issue where my generated Swift5 client included parameters of type `AnyCodable`, and I was getting the error: `Cannot find 'AnyCodable' in scope`. Adding `import AnyCodable` to the top of the generated file fixes the error, so I'm assuming that adding it to the template would fix the generated file.